students/K33422/Iskhakova_Emina/labs/2sem/requirements.txt
Django==3.2.10
django-allauth==0.47.0
django-bootstrap3==21.1
django-cors-headers==3.10.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-yasg==1.20.0
psycopg2==2.8.6
django_filter==21.1
python-magic==0.4.27
